Updates Maze.c
Recursive flood fill.
Reorganized Node and Maze structs.
The plan is now make a module to test the maze… aka make a driver to
test flood fill and add walls (manual or file).
Then if that works, load onto mouse once that is ready.
Stuff in main will be uncommented once i sort out the functions in
mylib.h (stuff for reading in input etc, which will not be needed for
actual micro mouse, just for testing this code).