Figured out the compiling error and...
DETAILS:
* For including multiple files to be compiled and loaded onto Maple,
you have to go to Sketched > Add file… which will show the other files.
For main_template, we need to have the Maze.c, Maze.h, Solver.c in the
same folder and included.
* Figured out why the compiler kept complaining about pointers… needed
typedef declaration.
* Putting everything into one maple .pde file does not work well with
function declarations near the top of file, because the pre processing
caused them to be placed at the very top… before the struct
definitions, which was why it kept complaining.